requirements.txt
Django==5.2
djangorestframework==3.15
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.3.1
coverage==7.12.0